Move tests for pyanaconda.ui.lib.users to a new file
Move the tests to a separate file.

Related: rhbz#2026916
